Cleanup windows warnings.

+++ include/message_filters/cache.h
@@ -208,13 +208,13 @@ public:
 
     std::lock_guard<std::mutex> lock(cache_lock_);
     // Find the starting index. (Find the first index after [or at] the start of the interval)
-    unsigned int start_index = cache_.size()-1;
+    size_t start_index = cache_.size() - 1;
     while(start_index > 0 &&
           mt::TimeStamp<M>::value(*cache_[start_index].getMessage()) > start)
     {
       start_index--;
     }
-    unsigned int end_index = start_index;
+    size_t end_index = start_index;
     while(end_index < cache_.size()-1 &&
           mt::TimeStamp<M>::value(*cache_[end_index].getMessage()) < end)
     {
@@ -223,7 +223,7 @@ public:
 
     std::vector<MConstPtr> interval_elems;
     interval_elems.reserve(end_index - start_index + 1) ;
-    for (unsigned int i=start_index; i<=end_index; i++)
+    for (size_t i=start_index; i<=end_index; i++)
     {
       interval_elems.push_back(cache_[i].getMessage()) ;
     }
@@ -272,12 +272,12 @@ public:
 
     MConstPtr out ;
 
-    int i=cache_.size()-1 ;
+    size_t i=cache_.size()-1 ;
     int elem_index = -1 ;
     while (i>=0 &&
         mt::TimeStamp<M>::value(*cache_[i].getMessage()) > time)
     {
-      elem_index = i ;
+      elem_index = static_cast<int>(i) ;
       i-- ;
     }
